% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_esu.R
\name{assign_esu}
\alias{assign_esu}
\title{Assign/create ESUs based on the sequences}
\usage{
assign_esu(path = "~/Documents/plankton-larvae-data/seqs", ...)
}
\arguments{
\item{path}{path indicating where the COI sequences are stored.}
}
\description{
Assign/create ESUs based on the sequences
}

