% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{generate_folder_structure}
\alias{generate_folder_structure}
\title{Creates the folder to store the photo in the archives}
\usage{

  generate_folder_structure(photo_folder = "/home/francois/hdd/plankton-images/archive_photos",
  sample_data = get_lab("sample_data"))
}
\arguments{
\item{photo_folder}{path where the photos are stored.}

\item{sample_data}{the data frame with the sample information}
}
\description{
Creates the folder to store the photo in the archives
}

