% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oligo_files.R
\name{generate_oligo_files}
\alias{generate_oligo_files}
\title{Function to generate the oligos files needed by Mothur for
de-multiplexing}
\usage{
generate_oligo_files(run = "BCS-003", data = get_lab("pcr_metacommunity"),
  primers = "leray", barcodes = get_lab("tags_primers"), outdir)
}
\arguments{
\item{run}{name of the sequencing run}

\item{data}{table with metadata info about the run}

\item{primers}{for now just leray for the usual COI leray primer
combination}

\item{barcodes}{the table with the correspondance between the
primer tag names and their respective sequences}

\item{outdir}{where the oligos files should be written}
}
\description{
Function to generate the oligos files needed by Mothur for
de-multiplexing
}
\details{
Currently the function output oligo files with the same
    naming scheme as returned by ICBR but it might not be the
    best in the long run.
}

