% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{generate_thumbnails}
\alias{generate_thumbnails}
\title{Creates the copies of the photos to be used by the shiny app}
\usage{

  generate_thumbnails(archive_folder = "/home/francois/hdd/plankton-images/archive_photos",
  app_folder = "/home/francois/hdd/plankton-images/app_photos",
  voucher_pattern = "FLPK-[0-9]{4}")
}
\arguments{
\item{archive_folder}{folder where the original photos are stored}

\item{app_folder}{folder where the smaller versions are stored}

\item{voucher_pattern}{regular expression that match the voucher}
}
\description{
Creates the copies of the photos to be used by the shiny app
}

