% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_lab}
\alias{get_lab}
\title{Load the data associated with the project}
\usage{
get_lab(what, path = "~/Documents/plankton-larvae-data", ...)
}
\arguments{
\item{what}{the name of the CSV file to load}

\item{path}{where the CSV files for the project live}

\item{...}{additional parameters to be passed to read.csv (stringsAsFactors is set to FALSE)}
}
\description{
Load the data associated with the project
}

