% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labmanager.R
\name{osf_render_template}
\alias{osf_render_template}
\title{Search an osf component for all modified subcomponents and files.}
\usage{
osf_render_template(data, template = osf_template$simple)
}
\arguments{
\item{data}{Either tibble returned from osf_retrieve/report_* functions, or list of meta data.}

\item{template}{Template to use for representing meta data. Passed to stringr::str_interp. See osf_template$simple for example.}
}
\value{
A character vector with descriptions for each entry.
}
\description{
Search an osf component for all modified subcomponents and files.
}
\examples{

osf_retrieve_node("yu8zm") \%>\% osf_render_template()

osf_retrieve_node("yu8zm") \%>\% osf_render_template("Name: ${attributes$title}")
}
