% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labmanager.R
\name{osf_report_modified}
\alias{osf_report_modified}
\title{Search an osf component for all modified subcomponents and files.}
\usage{
osf_report_modified(id, after_date, type = "user", template = NULL,
  ...)
}
\arguments{
\item{id}{Single osf_tbl row to search for subcomponents and files.}

\item{after_date}{Only keep components and files modified after this date.}

\item{type}{What type of id is used, "user" or "component".}

\item{template}{String template for formatting results.}

\item{...}{Additional arguments passed to walk_parent (only used if type is "component").}
}
\value{
A list of osf_tbl objects that can be combined using dplyr::bind_rows
}
\description{
Search an osf component for all modified subcomponents and files.
}
\examples{

me <- osf_retrieve_user("aswnc")

# return tibble of modified components and files
report <- osf_report_modified("aswnc", after_date = "2019-01-01")
report


# return printed reports only
osf_report_modified("aswnc", after_date = "2019-01-01", template = osf_template$simple)

# "6sqvw" is the id for the paper "Inducing cognitive control..."
osf_report_modified("6sqvw", type = "node", after_date = "2019-01-01")

# don't report files
osf_report_modified("6sqvw", type = "node", after_date = "2019-01-01")


}
