% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labmanager.R
\name{osf_report_modified_slack}
\alias{osf_report_modified_slack}
\title{Search an osf component for all modified subcomponents and files.}
\usage{
osf_report_modified_slack(id, after_date, type = "user",
  template = osf_template$simple, ...)
}
\arguments{
\item{id}{passed to osf_report_modified.}

\item{after_date}{passed to osf_report_modified.}

\item{type}{passed to osf_report_modified.}

\item{template}{passed to osf_report_modified.}

\item{...}{passed to slackr::slackr_bot

See slackr::slackr_setup for details on setting up slack credentials.}
}
\description{
Search an osf component for all modified subcomponents and files.
}
\examples{
## Not run:
# slackr_setup(incoming_webhook_url = "SOME_WEBHOOK_URL")
osf_report_modified_slack("aswnc", after_date = "2019-01-01", type = "user")

## End(Not run)
}
