% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-limsr.R
\name{run_limsr}
\alias{run_limsr}
\title{Run limsr}
\usage{
run_limsr()
}
\value{
Start the shiny app
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Francois Michonneau
}

