% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{uppercase_photo_extension}
\alias{uppercase_photo_extension}
\title{Check that the photo files have uppercase JPG extension}
\usage{

  uppercase_photo_extension(photo_folder = "/home/francois/hdd/plankton-images/archive_photos/")
}
\arguments{
\item{photo_folder}{path where the original photos are stored}
}
\description{
Check that the photo files have uppercase JPG extension
}

