% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labmanager.R
\name{walk_parent}
\alias{walk_parent}
\title{Search an osf component for all modified subcomponents and files.}
\usage{
walk_parent(parent_node, after_date, max_depth = 3, max_children = 30,
  sleep = 0.2, include_files = TRUE, level = 1)
}
\arguments{
\item{parent_node}{Single osf_tbl row to search for subcomponents and files.}

\item{after_date}{Only search osf subcomponents modified after this datetime.}

\item{max_depth}{Maximum depth of subcomponents to search.}

\item{max_children}{Don't search subcomponents that have more than max_children nodes.}

\item{sleep}{Time to wait before getting more subcomponents.}

\item{include_files}{Whether to search for changed files in addition to components.}

\item{level}{How deep the parent_node's children are (you shouldn't need this).}
}
\value{
A list of osf_tbl objects that can be combined using dplyr::bind_rows
}
\description{
Search an osf component for all modified subcomponents and files.
}
\examples{

}
